Frontend of Webshop
During this process, the entire frontend of the webshop app has bene developed. More changes may come in the future but for now, this is the main version of the website. All required webpages have been developed using HTML, React,js and styled components. Any further references will be present in the dissertation.

The next step is to develop the backend of the application and make it as functional as possible.