API Models and Routes created (Backend)
In this update, the api of the website has been developed although it has not been implemented into the website yet. Testing was done using Postman and the software I relied on to store the database and its data was MongoDB. The events that took place throughout this stage will be further explained in the dissertation.