Final version of Backend completed
In this commit, the API server and the UI of the webshop have been connected. Especially in the Product's page, where I have made the filters for product type work, and also made  each product selectable with its respective information in the same page. The Login and Signup pages were also refined, each field is required to be filled but unfortunately could not get the information put into those forms to be sent or fetched from the database. The remaining Backend (mainly for the Shopping Cart) could not be complete due to me running out of time.